modularizing code
OOP
