+++ src/backend/api_routes.py
@@ -1,5 +1,8 @@
 from flask import Blueprint, jsonify, render_template, request, session
-from gpt_service import generate_patient_case, load_diagnoses, generate_sbar_report, chat_with_patient, evaluate_diagnosis, assess_user_input
+from gpt_services.openai_client import client
+from gpt_services.case_generator import generate_patient_case, load_diagnoses, generate_sbar_report
+from gpt_services.chat_manager import chat_with_patient
+from gpt_services.evaluation import evaluate_diagnosis, assess_user_input
 
 api_routes = Blueprint('api', __name__)
 
@@ -16,11 +19,11 @@ def chat_init():
 json_file_path = '../../media/icd10_diagnoses.json'
 diagnoses = load_diagnoses(json_file_path)
 
-# Endpoint to generate a new patient case
+# Generate Case
 @api_routes.route('/generate-case', methods=['GET'])
 def generate_case():
     try:
-        case = generate_patient_case(diagnoses)
+        case = generate_patient_case(client, diagnoses)
         print(f"AI response: {case}")
 
         session['patient_case'] = case
@@ -31,7 +34,7 @@ def generate_case():
     except Exception as e:
         return jsonify({"error": str(e)}), 500
 
-# Endpoint for chat messages related to the patient case
+# Chat
 @api_routes.route('/chat/messages', methods=['POST'])
 def chat_message():
     user_input = request.json.get('user_input')
@@ -40,22 +43,18 @@ def chat_message():
     if not patient_case:
         return jsonify({"error": "No patient case found."}), 404
 
-    # Retrieve chat history from session
     chat_history = session.get('chat_history', [])
     
-    # Call the chat function with the user input, patient case, and current chat history
-    response = chat_with_patient(user_input, patient_case, chat_history).get('message')
+    response = chat_with_patient(client, user_input, patient_case, chat_history).get('message')
     print(f"AI response: {response}")
     
-    # Update chat history with user input and AI response
     chat_history.append(f"User: {user_input}")
     chat_history.append(f"Patient: {response}")
     
-    # Save updated chat history back into session
     session['chat_history'] = chat_history
     
-    if assess_user_input(user_input):  # Check if the user seems to give a diagnosis
-        evaluation = evaluate_diagnosis(patient_case, chat_history)
+    if assess_user_input(client, user_input):  # Check if the user seems to give a diagnosis
+        evaluation = evaluate_diagnosis(client, patient_case, chat_history)
         return jsonify({
             "response": response,
             "history": chat_history,
